[KYUUBI #1920] Update
